requirements.txt
bottle =>0.12.19
dominate =>2.3.1
pywebview =>3.3.5
PyQt5 =>12.9.1
